test_requirements.txt
pytest
pytest-django
psycopg2
